결제 후 상품 정보 저장 완료

+++ Project_test/src/main/java/com/example/demo/order/OrderDetailDto.java
@@ -19,9 +19,9 @@ public class OrderDetailDto {
     @NotNull
     private String image; // 이미지 URL
     @NotNull
-    private int detailEa; // 수량
+    private Long detailEa; // 수량
     @NotNull
-    private int price; // 가격
+    private Long price; // 가격
     @NotNull
     private String reviewWritten; // 리뷰 작성 여부 (Y/N)
